scripts/requirements-tests.txt
playwright>=1.40.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
